Add Environment Variables file to reducers.
This does contain all reducers and parsers for the
environment variables.
