+++ plugins/services/src/js/reducers/JSONConfigReducers.js
@@ -4,6 +4,9 @@ import {
   simpleReducer
 } from '../../../../../src/js/utils/ReducerUtil';
 import {
+  JSONReducer as env
+} from './form/EnvironmentVariables';
+import {
   SET,
   ADD_ROW,
   REMOVE_ROW
@@ -16,61 +19,7 @@ module.exports = {
   disk: simpleIntReducer('disk', 0),
   instances: simpleIntReducer('instances', 1),
   cmd: simpleReducer('cmd', ''),
-  env(state, {type, path, value}) {
-    if (this.env == null) {
-      this.env = [];
-
-      if (state != null) {
-        this.env = Object.keys(state).reduce((memo, key) => {
-          memo.push({
-            key: key.toUpperCase(),
-            value: state[key]
-          });
-          return memo;
-        }, []);
-      }
-    }
-
-    if (path != null && path.join('.').search('env') !== -1) {
-      if (path.join('.') === 'env') {
-        switch (type) {
-          case ADD_ROW:
-            this.env.push({key: null, value: null});
-            break;
-          case REMOVE_ROW:
-            this.env = this.env.filter((item, index) => {
-              return index !== value;
-            });
-            if (this.env.length === 0) {
-              this.env.push({key: null, value: null});
-            }
-            break;
-        }
-
-        return this.env.reduce((memo, item) => {
-          if (item.key != null || item.value != null) {
-            memo[item.key] = item.value;
-          }
-          return memo;
-        }, {});
-      }
-
-      const joinedPath = path.join('.');
-      const index = joinedPath.match(/\d+/)[0];
-      switch (true) {
-        case (type === SET && `env.${index}.key` === joinedPath):
-          this.env[index].key = value;
-          break;
-        case (type === SET && `env.${index}.value` === joinedPath):
-          this.env[index].value = value;
-          break;
-      }
-    }
-    return this.env.reduce((memo, item) => {
-      memo[item.key] = item.value;
-      return memo;
-    }, {});
-  },
+  env,
   labels(state, {type, path, value}) {
     if (this.labels == null) {
       this.labels = [];
